The canvas names an entity it has no number for by its # handle, as the CLI does (T-40493)
A reference the canvas had not loaded (the yak process's actor on "created
by") read as the bare uuid: formatProp fell back to the stored value, where
the CLI prints @yaks/id's `short`. Its fallback is now that handle.

The web kept its own copies of @yaks/id's handle, pattern, letter and id
(shortId, SHORT, prefixOf, idOf); they are gone, and every web door reads an
id through @yaks/id. An untitled entity's Inline chip is called by its id,
not its kind, and yak.sh v2's graph pane and boot log name an entity they no
longer hold by its handle rather than a slice or the whole uuid.


Tasks-Session: S-35798 981db89b-d3eb-49e8-9ece-047f6858b5f4

+++ packages/web/props.ts
@@ -9,6 +9,7 @@ import {
   spineProps,
   stamped,
 } from './types.ts'
+import { short } from '@yaks/id'
 import { timeInstant } from '@yaks/query'
 import { local } from './time.ts'
 import { normalize } from './url.ts'
@@ -243,8 +244,10 @@ export let formatProp = (
   if (tag(p.type) == 'bool') return parsed ? 'true' : 'false'
   // A stamp is stored Zulu but SHOWN local — one door for every face.
   if (tag(p.type) == 'time') return local(String(parsed))
+  // A reference nothing here can describe (its target not loaded) reads as
+  // the target's short handle, as the CLI prints it — never the bare uuid.
   if (tag(p.type) == 'eid') {
-    return ctx.describe?.(String(parsed)) ?? String(parsed)
+    return ctx.describe?.(String(parsed)) ?? short(String(parsed))
   }
   return String(parsed)
 }
